adding timestep scaling for r-value output

+++ src/main/java/uk/co/ramp/statistics/types/RValueOutput.java
@@ -13,7 +13,7 @@ import org.immutables.value.Value;
 @JsonPropertyOrder({"time", "newInfectors", "newInfections", "r", "sevenDayAverageR"})
 public interface RValueOutput {
 
-  int time();
+  double time();
 
   int newInfectors();
